refactor: update gateway/ln-gateway/src/lib.rs
Move contents of gateway/ln-gateway/src/gateway.rs to
gateway/ln-gateway/src/lib.rs
