+++ gateway/ln-gateway/src/bin/gatewayd.rs
@@ -13,8 +13,8 @@ use fedimint_core::task::TaskGroup;
 use fedimint_core::ServerModule;
 use fedimint_logging::TracingSetup;
 use ln_gateway::client::{DynGatewayClientBuilder, RocksDbFactory, StandardGatewayClientBuilder};
-use ln_gateway::gateway::Gateway;
 use ln_gateway::lnrpc_client::{DynLnRpcClient, NetworkLnRpcClient};
+use ln_gateway::Gateway;
 use mint_client::modules::ln::{Lightning, LightningGen};
 use mint_client::modules::mint::{Mint, MintGen};
 use mint_client::modules::wallet::{Wallet, WalletGen};
