+++ integrationtests/tests/fixtures/mod.rs
@@ -54,8 +54,8 @@ use hbbft::honey_badger::Batch;
 use itertools::Itertools;
 use ln_gateway::actor::GatewayActor;
 use ln_gateway::client::{DynGatewayClientBuilder, MemDbFactory, StandardGatewayClientBuilder};
-use ln_gateway::gateway::Gateway;
 use ln_gateway::lnrpc_client::{DynLnRpcClient, NetworkLnRpcClient};
+use ln_gateway::Gateway;
 use mint_client::mint::SpendableNote;
 use mint_client::transaction::legacy::Transaction;
 use mint_client::transaction::TransactionBuilder;
